Fix subtle performance issue with vectorisation & broadcasting
The issue is that to achieve broadcasting, we need a dynamic test in the
inner expression template loop. No way around I could find.

So let's add a preliminary check on the ET, and if the ET appear to
*not* need any broadcasting, then another vectorization loop, condition
free, is selected at runtime (which means both version are actually
compiled).